megre to evebox v0.3.2 also add org_members this removed the ability to delete your profile if you are a member of the company

+++ internal/core/modules/account/delete_account.go
@@ -13,6 +13,19 @@ func (s Service) DeleteOwnAccount(ctx context.Context, initiator InitiatorData) 
 		return err
 	}
 
+	exists, err := s.repo.ExistOrgMemberByAccount(ctx, initiator.AccountID)
+	if err != nil {
+		return errx.ErrorInternal.Raise(
+			fmt.Errorf("failed to check if account have membership in organizations, cause %s", err),
+		)
+	}
+
+	if exists {
+		return errx.AccountHaveMembershipInOrg.Raise(
+			fmt.Errorf("account %s has a member of organizations", initiator.AccountID),
+		)
+	}
+
 	return s.repo.Transaction(ctx, func(txCtx context.Context) error {
 		err = s.repo.DeleteAccount(ctx, initiator.AccountID)
 		if err != nil {
